gov_event_notive_monitor_v1.0/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cssselect==1.3.0
cycler==0.12.1
distro==1.9.0
docxcompose==1.4.0
docxtpl==0.19.1
exceptiongroup==1.2.2
fake-useragent==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Session==0.8.0
fonttools==4.56.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
kiwisolver==1.4.8
lxml==5.3.1
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgspec==0.19.0
networkx==3.4.2
numpy==2.1.1
openai==1.63.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parse==1.20.2
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.1
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
safetensors==0.5.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
Werkzeug==3.1.3
zipp==3.23.0
